Refactor branch creation & update (#74)
This commit refactors the way we create commits and branches to use the low-level GitHub "git data" API endpoints. in particular, it commits the file first and creates the branch second, rather than vice versa. This prevents unnecessary builds in CI due to the (empty) branch creation we had in previous versions. Closes #82.